+++ skeleton/TableType.tpl.php
@@ -3,12 +3,11 @@
 namespace <?= $namespace ?>;
 
 use <?= $entity->getFullName() ?>;
-<?php if ($entity_searchable) { ?>
 use Doctrine\ORM\QueryBuilder;
 use Umbrella\AdminBundle\Lib\Form\SearchType;
-<?php } ?>
 use Umbrella\AdminBundle\Lib\DataTable\Action\ButtonAddActionType;
 use Umbrella\AdminBundle\Lib\DataTable\Column\ActionColumnType;
+use Umbrella\AdminBundle\Lib\DataTable\Column\PropertyColumnType;
 use Umbrella\AdminBundle\Lib\DataTable\ColumnActionBuilder;
 use Umbrella\AdminBundle\Lib\DataTable\DataTableBuilder;
 use Umbrella\AdminBundle\Lib\DataTable\DataTableType;
@@ -17,9 +16,8 @@ class <?= $class_name ?> extends DataTableType
 {
     public function buildTable(DataTableBuilder $builder, array $options): void
     {
-<?php if ($entity_searchable) { ?>
         $builder->addFilter('search', SearchType::class);
-<?php } ?>
+
         $builder->addAction('add', ButtonAddActionType::class, [
             'route' => '<?= $route['name_prefix'] ?>_edit',
 <?php if ('modal' === $edit_view_type) { ?>
@@ -27,7 +25,9 @@ class <?= $class_name ?> extends DataTableType
 <?php } ?>
         ]);
 
-        $builder->add('id');
+        $builder->add('id', PropertyColumnType::class, [
+            'render' => fn(<?= $entity->getShortName() ?> $e) => sprintf('# %d', $e->id)
+        ]);
         $builder->add('__action__', ActionColumnType::class, [
             'build' => function (ColumnActionBuilder $builder, <?= $entity->getShortName() ?> $e) {
                 $builder->editLink([
@@ -46,14 +46,15 @@ class <?= $class_name ?> extends DataTableType
 
         $builder->useEntityAdapter([
             'class' => <?= $entity->getShortName() ?>::class,
-<?php if ($entity_searchable) { ?>
             'query' => function(QueryBuilder $qb, array $formData) {
+
+                // TODO : add closure to filter result depending on $formData['search'] value
                 if (isset($formData['search'])) {
-                    $qb->andWhere('LOWER(e.search) LIKE :search');
-                    $qb->setParameter('search', '%' . $formData['search'] . '%');
+                    // You can use Umbrella\AdminBundle\Utils\DoctrineUtils to filter results
+                    // DoctrineUtils::matchAll($qb, ['e.x', 'e.y'], $formData['search']);
                 }
+
             }
-<?php } ?>
         ]);
     }