+++ src/Lib/Search/Attribute/SearchableField.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Umbrella\AdminBundle\Lib\Search\Attribute;
-
-#[\Attribute(\Attribute::TARGET_METHOD | \Attribute::TARGET_PROPERTY)]
-class SearchableField
-{
-}